backend/config.py
"""BioMentor AI — Configuration"""
import os
from dotenv import load_dotenv

load_dotenv()

# ── LLM ──────────────────────────────────────────────
GROQ_API_KEY = os.getenv("GROQ_API_KEY", "")
LLM_MODEL = "llama-3.3-70b-versatile"
LLM_TEMPERATURES = {
    "content": 0.5,
    "course": 0.5,
    "quiz": 0.3,
    "evaluation": 0.2,
    "adaptive": 0.2,
    "chat": 0.4,
}
LLM_MAX_TOKENS = 4096

# ── Paths ────────────────────────────────────────────
BASE_DIR = os.path.dirname(os.path.abspath(__file__))
DATA_DIR = os.path.join(BASE_DIR, "data")
CHROMA_DIR = os.path.join(DATA_DIR, "chroma_db")
UPLOADS_DIR = os.path.join(DATA_DIR, "uploads")
DB_PATH = os.path.join(DATA_DIR, "biomentor.db")

# ── Embedding ────────────────────────────────────────
EMBEDDING_MODEL = "all-MiniLM-L6-v2"

# ── RAG ──────────────────────────────────────────────
CHUNK_SIZE = 1000
CHUNK_OVERLAP = 200
TOP_K = 5

# Ensure directories exist
os.makedirs(DATA_DIR, exist_ok=True)
os.makedirs(CHROMA_DIR, exist_ok=True)
os.makedirs(UPLOADS_DIR, exist_ok=True)
