backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# LLM
langchain==0.3.14
langchain-groq==0.2.4
langchain-community==0.3.14

# RAG
chromadb==0.5.23
sentence-transformers==3.3.1
pypdf2==3.0.1

# Database
aiosqlite==0.20.0

# Utilities
python-dotenv==1.0.1
pydantic==2.10.4
